# perf: optimize `ReadFrame`